requirements.txt
python-dotenv==1.0.1
requests==2.32.3
openai==1.52.2